Dynamic collections

+++ es-app/src/guis/GuiGamelistFilter.cpp
@@ -8,6 +8,13 @@
 
 GuiGamelistFilter::GuiGamelistFilter(Window* window, SystemData* system) : GuiComponent(window), mMenu(window, _("FILTER GAMELIST BY")), mSystem(system)
 {
+	mFilterIndex = mSystem->getIndex(true);
+	initializeMenu();
+}
+
+GuiGamelistFilter::GuiGamelistFilter(Window* window, FileFilterIndex* filterIndex) : GuiComponent(window), mMenu(window, _("EDIT COLLECTION FILTERS")), mSystem(nullptr)
+{
+	mFilterIndex = filterIndex;
 	initializeMenu();
 }
 
@@ -15,15 +22,16 @@ void GuiGamelistFilter::initializeMenu()
 {
 	addChild(&mMenu);
 
-	// get filters from system
-
-	mFilterIndex = mSystem->getIndex(true);
-
 	ComponentListRow row;
 
-	mMenu.addEntry(_("RESET ALL FILTERS"), false, std::bind(&GuiGamelistFilter::resetAllFilters, this));
+	if (mSystem != nullptr)
+		mMenu.addEntry(_("RESET ALL FILTERS"), false, std::bind(&GuiGamelistFilter::resetAllFilters, this));
+	else
+	{
+		addTextFilterToMenu();
+		addSystemFilterToMenu();
+	}
 
-	// addTextFilterToMenu();
 	addFiltersToMenu();
 
 	mMenu.addButton(_("BACK"), "back", std::bind(&GuiGamelistFilter::applyFilters, this));
@@ -37,7 +45,8 @@ void GuiGamelistFilter::initializeMenu()
 void GuiGamelistFilter::resetAllFilters()
 {
 	mFilterIndex->resetFilters();
-	for (std::map<FilterIndexType, std::shared_ptr< OptionListComponent<std::string> >>::const_iterator it = mFilterOptions.cbegin(); it != mFilterOptions.cend(); ++it ) {
+	for (std::map<FilterIndexType, std::shared_ptr< OptionListComponent<std::string> >>::const_iterator it = mFilterOptions.cbegin(); it != mFilterOptions.cend(); ++it ) 
+	{
 		std::shared_ptr< OptionListComponent<std::string> > optionList = it->second;
 		optionList->selectNone();
 	}
@@ -47,7 +56,7 @@ GuiGamelistFilter::~GuiGamelistFilter()
 {
 	mFilterOptions.clear();
 
-	if (!mFilterIndex->isFiltered())
+	if (!mFilterIndex->isFiltered() && mSystem != nullptr)
 		mSystem->deleteIndex();
 }
 
@@ -119,23 +128,6 @@ void GuiGamelistFilter::addFiltersToMenu()
 	}
 }
 
-void GuiGamelistFilter::applyFilters()
-{
-	if (mTextFilter)
-		mFilterIndex->setTextFilter(mTextFilter->getValue());
-
-	std::vector<FilterDataDecl> decls = mFilterIndex->getFilterDataDecls();
-	for (std::map<FilterIndexType, std::shared_ptr< OptionListComponent<std::string> >>::const_iterator it = mFilterOptions.cbegin(); it != mFilterOptions.cend(); ++it ) 
-	{
-		std::shared_ptr< OptionListComponent<std::string> > optionList = it->second;
-		std::vector<std::string> filters = optionList->getSelectedObjects();
-		mFilterIndex->setFilter(it->first, &filters);
-	}
-
-	delete this;
-
-}
-
 bool GuiGamelistFilter::input(InputConfig* config, Input input)
 {
 	bool consumed = GuiComponent::input(config, input);
@@ -153,4 +145,63 @@ std::vector<HelpPrompt> GuiGamelistFilter::getHelpPrompts()
 	std::vector<HelpPrompt> prompts = mMenu.getHelpPrompts();
 	prompts.push_back(HelpPrompt(BUTTON_BACK, _("BACK")));
 	return prompts;
+}
+
+void GuiGamelistFilter::addSystemFilterToMenu()
+{	
+	CollectionFilter* cf = dynamic_cast<CollectionFilter*>(mFilterIndex);
+	if (cf == nullptr)
+		return;
+	
+	mSystemFilter = std::make_shared<OptionListComponent<SystemData*>>(mWindow, _("SYSTEMS"), true);
+	
+	for (auto system : SystemData::sSystemVector)
+		if (!system->isCollection() && !system->isGroupSystem())
+			mSystemFilter->add(system->getFullName(), system, cf->isSystemSelected(system->getName()));
+
+	mMenu.addWithLabel(_("SYSTEMS"), mSystemFilter);
+
+
+}
+
+void GuiGamelistFilter::applyFilters()
+{
+	CollectionFilter* collectionFilter = dynamic_cast<CollectionFilter*>(mFilterIndex);
+
+	if (mTextFilter)
+		mFilterIndex->setTextFilter(mTextFilter->getValue());
+
+	if (mSystemFilter != nullptr && collectionFilter != nullptr)
+	{
+		std::string hiddenSystems;
+
+		std::vector<SystemData*> sys = mSystemFilter->getSelectedObjects();
+
+		if (!mSystemFilter->hasSelection() || mSystemFilter->size() == sys.size())
+			collectionFilter->resetSystemFilter();
+		else
+		{
+			for (auto system : SystemData::sSystemVector)
+			{
+				if (system->isCollection() || system->isGroupSystem())
+					continue;
+
+				bool sel = std::find(sys.cbegin(), sys.cend(), system) != sys.cend();
+				collectionFilter->setSystemSelected(system->getName(), sel);
+			}
+		}
+	}
+
+	std::vector<FilterDataDecl> decls = mFilterIndex->getFilterDataDecls();
+	for (std::map<FilterIndexType, std::shared_ptr< OptionListComponent<std::string> >>::const_iterator it = mFilterOptions.cbegin(); it != mFilterOptions.cend(); ++it)
+	{
+		std::shared_ptr< OptionListComponent<std::string> > optionList = it->second;
+		std::vector<std::string> filters = optionList->getSelectedObjects();
+		mFilterIndex->setFilter(it->first, &filters);
+	}
+
+	if (collectionFilter != nullptr)
+		collectionFilter->save();
+
+	delete this;
 }
